+++ KopiLuaMod/Lua/src/LuacProgram.cs
@@ -180,7 +180,7 @@ namespace KopiLua
 				f.source = LuaString.luaS_newliteral(L, CharPtr.toCharPtr("=(" + PROGNAME + ")"));
 				f.maxstacksize = 1;
 				pc = 2 * n + 1;
-				f.code = (UInt32[]/*Instruction[]*/)LuaMem.luaM_newvector<UInt32/*Instruction*/>(L, pc);
+				f.code = (long[]/*UInt32[]*//*Instruction[]*/)LuaMem.luaM_newvector<long/*UInt32*//*Instruction*/>(L, pc);
 				f.sizecode = pc;
 				f.p = LuaMem.luaM_newvector<Proto>(L, n);
 				f.sizep = n;
@@ -188,10 +188,10 @@ namespace KopiLua
 				for (i = 0; i < n; i++)
 				{
 					f.p[i] = toproto(L,i-n-1);
-					f.code[pc++] = (uint)LuaOpCodes.CREATE_ABx(OpCode.OP_CLOSURE, 0, i);
-					f.code[pc++] = (uint)LuaOpCodes.CREATE_ABC(OpCode.OP_CALL, 0, 1, 1);
+					f.code[pc++] = (long/*uint*/)LuaOpCodes.CREATE_ABx(OpCode.OP_CLOSURE, 0, i);
+					f.code[pc++] = (long/*uint*/)LuaOpCodes.CREATE_ABC(OpCode.OP_CALL, 0, 1, 1);
 				}
-				f.code[pc++] = (uint)LuaOpCodes.CREATE_ABC(OpCode.OP_RETURN, 0, 1, 0);
+				f.code[pc++] = (long/*uint*/)LuaOpCodes.CREATE_ABC(OpCode.OP_RETURN, 0, 1, 0);
 				return f;
 			}
 		}
